requirements.txt
clearbit==0.1.7
django==3.1.2
pyJWT==1.7.1
requests==2.24.0